refactor(catalog): move iceberg_rest_service + segment_registry into proximadb-catalog (root-crate decomp Slice 4)
Unblocks the two storage-common-coupled catalog files by moving the
cycle-causing contracts DOWN instead of hoisting a foundation crate:

- ObjectStoreBridge (+ CommitOutcome, IcebergSnapshotField, Bridge* aliases)
  descends from proximadb-storage-common into proximadb-catalog — the
  cross-layer contract crate storage-common already depends on; the trait's
  ProximaSchema param is the in-crate CatalogTableSchema. storage-common
  re-exports the module at the old path (zero reroutes for all users).
- SegmentMeta descends from storage-common pax_block into
  proximadb-block-format, next to the BlockStats it aggregates; pax_block
  re-exports it at the old path.
- src/catalog/{iceberg_rest_service,segment_registry}.rs (~1.7K lines) move
  into proximadb-catalog; root src/catalog/mod.rs keeps re-export shims so
  crate::catalog::* paths are unchanged.
- proximadb-iceberg-engine becomes a dev-dependency of the catalog crate
  (Cargo-legal dev cycle) for the bridge integration test, which moves from
  the lib #[cfg(test)] module to tests/iceberg_rest_service_bridge.rs:
  inside a unit test, crate:: trait identity differs from the
  proximadb_catalog the engine implements against.
- CAPABILITY_MATRIX.toml evidence path updated; decomposition SOT doc status
  section updated (the planned foundation hoist of ProximaSchema was
  infeasible: its field-type closure is ~the whole catalog contract surface).

Verified: crate+root clippy --lib --bins(-tests) -D warnings; fmt; workspace
boundaries clean with the storage up-edge ratchet held at 197; catalog (441)
+ storage-common (410) + block-format nextest suites green; server binary
builds.
