+++ crates/control/proximadb-catalog/Cargo.toml
@@ -44,11 +44,19 @@ proximadb-data-model = { workspace = true }
 proximadb-distance-types = { workspace = true }
 proximadb-quantization-types = { workspace = true }
 proximadb-records = { workspace = true }
+# The canonical `ObjectStoreBridge` contract module (moved down from
+# `proximadb-storage-common`, which re-exports it): object_store/arrow-array/
+# futures for the trait surface, kernel for `StorageError`.
+arrow-array = { workspace = true, default-features = false }
+futures = { workspace = true }
+object_store = { workspace = true }
+proximadb-kernel = { workspace = true }
+# `SegmentMeta` for the segment registry (block-format is a foundation-clean
+# leaf: serde + codec/data-model/records/filter-expression only — no cycle).
+proximadb-block-format = { workspace = true }
 # Foundation deps pulled in by the runtime/wiring modules extracted from the
 # root `src/catalog` (Slice 1): proto messages (Collection), filter expressions
-# (partition pruning), and the cache façade. (segment_registry stays root — it
-# needs storage_common::pax_block::SegmentMeta, and storage-common→catalog is a
-# downward edge, so catalog→storage-common would cycle.)
+# (partition pruning), and the cache façade.
 proximadb-cache = { workspace = true }
 proximadb-filter-expression = { workspace = true }
 proximadb-proto = { workspace = true }
@@ -65,5 +73,9 @@ tracing.workspace = true
 uuid = { version = "1.0", features = ["v4"] }  # opaque server-issued namespace_id (ns_<uuid>)
 
 [dev-dependencies]
+# Concrete bridge impl for the Iceberg REST service tests (dev-only cycle:
+# iceberg-engine sits above this crate in the main graph, which Cargo permits
+# for dev-dependencies).
+proximadb-iceberg-engine = { workspace = true }
 tempfile = { workspace = true }
 tokio = { version = "1.37", features = ["full", "test-util"] }
